requirements.txt
Flask==3.0.3
Flask-Session==0.5.0
firebase-admin==7.1.1
python-dotenv==1.0.1
pandas==2.2.2
scikit-learn==1.5.0
joblib==1.4.2
Babel==2.15.1
Werkzeug==3.1.4